Le MVC ça fout la merde

+++ Modele/suprcom.php
@@ -3,7 +3,10 @@ session_start();
 $id_utilisateur=$_SESSION['id_utilisateur'];
 
 $connect_e = mysqli_connect("localhost", "root", "", "bddsimplevent");
+if (!$connect_e) {
+    printf("Echec de la connexion : %s\n", mysqli_connect_error());
+    exit();
+	}
 mysqli_query($connect_e,"DELETE FROM commente WHERE id_commentaire=".$_GET['i_com']) ;
-header("Refresh:0 ,url=".$_SERVER['HTTP_REFERER']);
 
 ?>
